@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "vapor-steam"
-version = "1.5.8"
+version = "1.6.0"
 description = "TUI program to check the ProtonDB compatibility of all the games of a Steam user."
 authors = ["TabulateJarl8 <tabulatejarl8@gmail.com>"]
 license = "GPLv3"
@@ -16,7 +16,6 @@ classifiers = [
 	"License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
 	"Natural Language :: English",
 	"Operating System :: OS Independent",
-	"Programming Language :: Python :: 3.8",
 	"Programming Language :: Python :: 3.9",
 	"Programming Language :: Python :: 3.10",
 	"Programming Language :: Python :: 3.11",
@@ -27,19 +26,19 @@ classifiers = [
 include = [{ path = "tests", format = "sdist" }]
 
 [tool.poetry.dependencies]
-python = "^3.8.1"
-textual = "^0.79.1"
-aiohttp = "^3.9.5"
-platformdirs = "^4.3.2"
+python = "^3.9.0"
+textual = "^2.1.0"
+aiohttp = "^3.11.12"
+platformdirs = "^4.3.6"
 typing-extensions = "^4.12.2"
 
 [tool.poetry.scripts]
 vapor = "vapor.__main__:main"
 
 [tool.poetry.group.dev.dependencies]
-textual-dev = "^1.6.1"
-ruff = "^0.6.4"
-pytest = "^8.3.3"
+textual-dev = "^1.7.0"
+ruff = "^0.9.7"
+pytest = "^8.3.4"
 pytest-cov = "^5.0.0"
 pytest-asyncio = "^0.24.0"
 
@@ -72,18 +71,13 @@ exclude_also = [
 
 	# dont test functions that only make network calls
 	"async def async_get",
-	"async def get_anti_cheat_data",
-	"async def get_game_average_rating",
-	"async def resolve_vanity_name",
-	"async def get_steam_user_data",
-
-	# dont test status checking
-	"if data.status != 200",
-
 ]
 
 omit = ["*/__main__.py"]
 
+
+[tool.ruff]
+target-version = "py39"
 
 [tool.ruff.lint]
 preview = true
