@@ -1,6 +1,7 @@
 """Vapor UI tests."""
 
-from typing import Optional
+from __future__ import annotations
+
 from unittest.mock import Mock, patch
 
 import pytest
@@ -45,7 +46,7 @@ def config() -> Config:
 class MockCache:
 	"""Mock Cache object with set anticheat data."""
 
-	def get_anticheat_data(self, app_id: str) -> Optional[AntiCheatData]:
+	def get_anticheat_data(self, app_id: str) -> AntiCheatData | None:
 		"""Return anticheat status denied for id 123."""
 		if app_id == '123':
 			return AntiCheatData('123', AntiCheatStatus.DENIED)
@@ -125,9 +126,12 @@ def test_create_app() -> None:
 @pytest.mark.asyncio
 async def test_table_population_username(config: Config) -> None:
 	"""Test that table is populated correctly on submission."""
-	with patch('vapor.main.get_anti_cheat_data', return_value=MockCache()), patch(
-		'vapor.main.get_steam_user_data',
-		return_value=STEAM_USER_DATA,
+	with (
+		patch('vapor.main.get_anti_cheat_data', return_value=MockCache()),
+		patch(
+			'vapor.main.get_steam_user_data',
+			return_value=STEAM_USER_DATA,
+		),
 	):
 		app = SteamApp(config)
 
@@ -164,9 +168,12 @@ async def test_table_population_username(config: Config) -> None:
 @pytest.mark.asyncio
 async def test_parse_steam_url_id(config: Config) -> None:
 	"""Test that Steam URLs (/id/) are correctly parsed."""
-	with patch('vapor.main.get_anti_cheat_data', return_value=MockCache()), patch(
-		'vapor.main.get_steam_user_data',
-		return_value=STEAM_USER_DATA,
+	with (
+		patch('vapor.main.get_anti_cheat_data', return_value=MockCache()),
+		patch(
+			'vapor.main.get_steam_user_data',
+			return_value=STEAM_USER_DATA,
+		),
 	):
 		app = SteamApp(config)
 
@@ -185,9 +192,12 @@ async def test_parse_steam_url_id(config: Config) -> None:
 @pytest.mark.asyncio
 async def test_parse_steam_url_profiles(config: Config) -> None:
 	"""Test that Steam URLs (/profiles/) are correctly parsed."""
-	with patch('vapor.main.get_anti_cheat_data', return_value=MockCache()), patch(
-		'vapor.main.get_steam_user_data',
-		return_value=STEAM_USER_DATA,
+	with (
+		patch('vapor.main.get_anti_cheat_data', return_value=MockCache()),
+		patch(
+			'vapor.main.get_steam_user_data',
+			return_value=STEAM_USER_DATA,
+		),
 	):
 		app = SteamApp(config)
 
@@ -207,9 +217,12 @@ async def test_parse_steam_url_profiles(config: Config) -> None:
 @pytest.mark.asyncio
 async def test_no_cache_user_query(config: Config) -> None:
 	"""Test that anticheat data is not present without cache."""
-	with patch('vapor.main.get_anti_cheat_data', return_value=None), patch(
-		'vapor.main.get_steam_user_data',
-		return_value=STEAM_USER_DATA,
+	with (
+		patch('vapor.main.get_anti_cheat_data', return_value=None),
+		patch(
+			'vapor.main.get_steam_user_data',
+			return_value=STEAM_USER_DATA,
+		),
 	):
 		app = SteamApp(config)
 
@@ -224,9 +237,12 @@ async def test_no_cache_user_query(config: Config) -> None:
 @pytest.mark.asyncio
 async def test_user_id_preservation(config: Config) -> None:
 	"""Test that user ID is preserved when the setting is on."""
-	with patch('vapor.main.get_anti_cheat_data', return_value=None), patch(
-		'vapor.main.get_steam_user_data',
-		return_value=STEAM_USER_DATA,
+	with (
+		patch('vapor.main.get_anti_cheat_data', return_value=None),
+		patch(
+			'vapor.main.get_steam_user_data',
+			return_value=STEAM_USER_DATA,
+		),
 	):
 		app = SteamApp(config)
 
@@ -251,9 +267,12 @@ async def test_user_id_preservation(config: Config) -> None:
 @pytest.mark.asyncio
 async def test_invalid_id_error(config: Config) -> None:
 	"""Test that an error is appropriately displayed when an invalid ID is entered."""
-	with patch('vapor.main.get_anti_cheat_data', return_value=None), patch(
-		'vapor.main.get_steam_user_data',
-		side_effect=Mock(side_effect=InvalidIDError),
+	with (
+		patch('vapor.main.get_anti_cheat_data', return_value=None),
+		patch(
+			'vapor.main.get_steam_user_data',
+			side_effect=Mock(side_effect=InvalidIDError),
+		),
 	):
 		app = SteamApp(config)
 
@@ -268,9 +287,12 @@ async def test_invalid_id_error(config: Config) -> None:
 @pytest.mark.asyncio
 async def test_unauthorized_error(config: Config) -> None:
 	"""Test that invalid Steam API keys display the appropriate error."""
-	with patch('vapor.main.get_anti_cheat_data', return_value=None), patch(
-		'vapor.main.get_steam_user_data',
-		side_effect=Mock(side_effect=UnauthorizedError),
+	with (
+		patch('vapor.main.get_anti_cheat_data', return_value=None),
+		patch(
+			'vapor.main.get_steam_user_data',
+			side_effect=Mock(side_effect=UnauthorizedError),
+		),
 	):
 		app = SteamApp(config)
 
@@ -285,9 +307,12 @@ async def test_unauthorized_error(config: Config) -> None:
 @pytest.mark.asyncio
 async def test_private_account_screen(config: Config) -> None:
 	"""Test that the private account error screen shows."""
-	with patch('vapor.main.get_anti_cheat_data', return_value=None), patch(
-		'vapor.main.get_steam_user_data',
-		side_effect=Mock(side_effect=PrivateAccountError),
+	with (
+		patch('vapor.main.get_anti_cheat_data', return_value=None),
+		patch(
+			'vapor.main.get_steam_user_data',
+			side_effect=Mock(side_effect=PrivateAccountError),
+		),
 	):
 		app = SteamApp(config)
 
