@@ -1,7 +1,9 @@
 """Main code and UI."""
 
+from __future__ import annotations
+
 from pathlib import Path
-from typing import ClassVar, List, Optional, Union, cast
+from typing import ClassVar, cast
 from urllib.parse import urlparse
 
 from rich.text import Text
@@ -43,7 +45,7 @@ from vapor.exceptions import InvalidIDError, PrivateAccountError, UnauthorizedEr
 class SettingsScreen(Screen[None]):
 	"""Settings editor screen for modifying the config file."""
 
-	BINDINGS: ClassVar[List[BindingType]] = [
+	BINDINGS: ClassVar[list[BindingType]] = [
 		Binding('escape', 'app.pop_screen', 'Close Settings', show=True),
 	]
 
@@ -106,13 +108,13 @@ class PrivateAccountScreen(ModalScreen[None]):
 class SteamApp(App[None]):
 	"""Main application class."""
 
-	CSS_PATH: ClassVar[Union[CSSPathType, None]] = 'main.tcss'
-	TITLE: Union[str, None] = 'Steam Profile Proton Compatibility Checker'
-	BINDINGS: ClassVar[List[BindingType]] = [
+	CSS_PATH: ClassVar[CSSPathType | None] = 'main.tcss'
+	TITLE: str | None = 'Steam Profile Proton Compatibility Checker'
+	BINDINGS: ClassVar[list[BindingType]] = [
 		Binding('ctrl+s', "push_screen('settings')", 'Settings', show=True),
 	]
 
-	def __init__(self, custom_config: Optional[Config] = None) -> None:
+	def __init__(self, custom_config: Config | None = None) -> None:
 		"""Construct the application.
 
 		This reads and instantiates the config.
@@ -153,7 +155,7 @@ class SteamApp(App[None]):
 					id='user-rating',
 				),
 			),
-			DataTable[Union[str, Text]](zebra_stripes=True),
+			DataTable[str | Text](zebra_stripes=True),
 			id='body',
 		)
 		yield Footer()
@@ -161,7 +163,7 @@ class SteamApp(App[None]):
 	def on_mount(self) -> None:
 		"""On mount, we initialize the table columns."""
 		# add nothing to table so that it shows up
-		table: DataTable[Union[str, Text]] = self.query_one(DataTable)
+		table: DataTable[str | Text] = self.query_one(DataTable)
 		table.add_columns('Title', 'Compatibility', 'Anti-Cheat Compatibility')
 
 		for _ in range(12):
@@ -188,8 +190,8 @@ class SteamApp(App[None]):
 				item.refresh()
 
 			# set the DataTable as loading
-			table: DataTable[Union[str, Text]] = self.query_one(DataTable)
-			table.set_loading(loading=True)  # pyright: ignore[reportUnknownMemberType]
+			table: DataTable[str | Text] = self.query_one(DataTable)
+			table.set_loading(loading=True)
 
 			# get user's API key and ID
 			api_key: Input = cast(Input, self.query_one('#api-key'))
@@ -264,7 +266,7 @@ class SteamApp(App[None]):
 
 			# set table as not loading
 			table = self.query_one(DataTable)
-			table.set_loading(loading=False)  # pyright: ignore[reportUnknownMemberType]
+			table.set_loading(loading=False)
 
 			if self.show_account_help_dialog:
 				self.show_account_help_dialog = False
